Subiso adaptations for Norwegian.
