+++ ts/l10n/numbers/numbers_nn.ts
@@ -18,6 +18,7 @@
  * @author volker.sorge@gmail.com (Volker Sorge)
  */
 
+import Engine from '../../common/engine';
 import { Numbers, NUMBERS as NUMB } from '../messages';
 
 /**
@@ -27,7 +28,7 @@ import { Numbers, NUMBERS as NUMB } from '../messages';
  * @param ordinal Flag indicating if we construct an ordinal.
  * @returns The string representation of that number.
  */
-function hundredsToWords_(num: number, ordinal = false): string {
+function hundredsToWordsRo_(num: number, ordinal = false): string {
   let n = num % 1000;
   let str = '';
   const count = Math.floor(n / 100);
@@ -63,9 +64,9 @@ function hundredsToWords_(num: number, ordinal = false): string {
  * @param ordinal Flag indicating if we construct an ordinal.
  * @returns The string representation of that number.
  */
-function numberToWords(num: number, ordinal = false): string {
+function numberToWordsRo(num: number, ordinal = false): string {
   if (num === 0) {
-    return NUMBERS.zero;
+    return ordinal ? NUMBERS.special.smallOrdinals[0] : NUMBERS.zero;
   }
   if (num >= Math.pow(10, 36)) {
     return num.toString();
@@ -75,7 +76,7 @@ function numberToWords(num: number, ordinal = false): string {
   while (num > 0) {
     const hundreds = num % 1000;
     if (hundreds) {
-      const hund = hundredsToWords_(num % 1000, pos ? false : ordinal);
+      const hund = hundredsToWordsRo_(num % 1000, pos ? false : ordinal);
       if (!pos && ordinal) {
         ordinal = !ordinal;
       }
@@ -161,3 +162,102 @@ NUMBERS.numberToOrdinal = numberToOrdinal;
 
 export default NUMBERS;
 // TODO: For simple speech output this should be different.
+
+// Alternative Germanic style numbers:
+/**
+ * Changes number one 'eins' into a prefix.
+ *
+ * @param num Number string.
+ * @param thd Flag indicating if this a thousand or above.
+ * @returns If it is a one, it is made into prefix.
+ */
+function onePrefix_(num: string, thd = false): string {
+  const numOne = NUMBERS.ones[1];
+  return num === numOne ? (num === 'ein' ? 'eitt ' : thd ? 'et' : 'ett') : num;
+}
+
+/**
+ * Translates a number of up to twelve digits into a string representation.
+ *
+ * @param num The number to translate.
+ * @param ordinal Flag indicating if we construct an ordinal.
+ * @returns The string representation of that number.
+ */
+function hundredsToWordsGe_(num: number, ordinal = false): string {
+  let n = num % 1000;
+  let str = '';
+  let ones = NUMBERS.ones[Math.floor(n / 100)];
+  str += ones ? onePrefix_(ones) + 'hundre' : '';
+  n = n % 100;
+  if (n) {
+    str += str ? 'og' : '';
+    if (ordinal) {
+      const ord = NUMBERS.special.smallOrdinals[n];
+      if (ord) {
+        return str += ord;
+      }
+    }
+    ones = NUMBERS.ones[n];
+    if (ones) {
+      str += ones;
+    } else {
+      const tens = NUMBERS.tens[Math.floor(n / 10)];
+      ones = NUMBERS.ones[n % 10];
+      str += ones ? ones + 'og' + tens : tens;
+    }
+  }
+  return ordinal ? replaceOrdinal(str) : str;
+}
+
+/**
+ * Translates a number of up to twelve digits into a string representation.
+ *
+ * @param num The number to translate.
+ * @param ordinal Flag indicating if we construct an ordinal.
+ * @returns The string representation of that number.
+ */
+function numberToWordsGe(num: number, ordinal = false): string {
+  if (num === 0) {
+    return ordinal ? NUMBERS.special.smallOrdinals[0] : NUMBERS.zero;
+  }
+  if (num >= Math.pow(10, 36)) {
+    return num.toString();
+  }
+  let pos = 0;
+  let str = '';
+  while (num > 0) {
+    const hundreds = num % 1000;
+    if (hundreds) {
+      const hund = hundredsToWordsGe_(num % 1000, pos ? false : ordinal);
+      if (!pos && ordinal) {
+        ordinal = !ordinal;
+      }
+      str =
+        (pos === 1 ? onePrefix_(hund, true) : hund) +
+        (pos > 1 ? NUMBERS.numSep : '') +
+        (pos
+          ? // If this is million or above take care oaf the plural.
+            NUMBERS.large[pos] + (pos > 1 && hundreds > 1 ? 'er' : '')
+          : '') +
+        (pos > 1 && str ? NUMBERS.numSep : '') +
+        str;
+    }
+    num = Math.floor(num / 1000);
+    pos++;
+  }
+  return ordinal ? str + (str.match(/tusen$/) ? 'de' : 'te') : str;
+}
+
+/**
+ * Translates a number of up to twelve digits into a string representation.
+ *
+ * @param num The number to translate.
+ * @param ordinal Flag indicating if we construct an ordinal.
+ * @returns The string representation of that number.
+ */
+function numberToWords(num: number, ordinal = false): string {
+  const word = Engine.getInstance().subiso === 'alt' ?
+    numberToWordsGe(num, ordinal) :
+    numberToWordsRo(num, ordinal);
+  return word;
+}
